Various fixes for partner permissions:
Correctly filter project lists
Correctly calculate related OneOf permission
Add missing skills permission for authenticated users
Fix wallpost filtering and permissions
