+++ bluebottle/bb_projects/views.py
@@ -1,6 +1,11 @@
+from dateutil import parser
+import datetime
 from django.db.models.query_utils import Q
 
+from django.db.models import F
+from django.db.models.aggregates import Count
 from django.contrib.auth.models import AnonymousUser
+from django.utils import timezone
 
 from bluebottle.projects.models import Project, ProjectPhaseLog, ProjectDocument
 from bluebottle.bluebottle_drf2.pagination import BluebottlePagination
@@ -28,30 +33,142 @@ class TinyProjectPagination(BluebottlePagination):
     page_size = 10000
 
 
-class ProjectTinyPreviewList(OwnerListViewMixin, ListAPIView):
+class ProjectListSearchMixin(object):
+
+    def search(self, qs, query):
+        # Apply filters
+        status = query.getlist(u'status[]', None)
+        if status:
+            qs = qs.filter(status__slug__in=status)
+        else:
+            status = query.get('status', None)
+            if status:
+                qs = qs.filter(status__slug=status)
+
+        country = query.get('country', None)
+        if country:
+            qs = qs.filter(country=country)
+
+        location = query.get('location', None)
+        if location:
+            qs = qs.filter(location=location)
+
+        category = query.get('category', None)
+        if category:
+            qs = qs.filter(categories__slug=category)
+
+        theme = query.get('theme', None)
+        if theme:
+            qs = qs.filter(theme_id=theme)
+
+        money_needed = query.get('money_needed', None)
+        if money_needed:
+            qs = qs.filter(amount_needed__gt=0)
+
+        skill = query.get('skill', None)
+        if skill:
+            qs.select_related('task')
+            qs = qs.filter(task__skill=skill).distinct()
+
+        anywhere = query.get('anywhere', None)
+        if anywhere:
+            qs = qs.filter(Q(task__id__isnull=False), Q(task__location__isnull=True) | Q(task__location='')).distinct()
+
+        start = query.get('start', None)
+        if start:
+            qs.select_related('task')
+
+            tz = timezone.get_current_timezone()
+            start_date = tz.localize(
+                datetime.datetime.combine(parser.parse(start), datetime.datetime.min.time())
+            )
+
+            end = query.get('end', start)
+            end_date = tz.localize(
+                datetime.datetime.combine(parser.parse(end), datetime.datetime.max.time())
+            )
+
+            qs = qs.filter(
+                Q(task__type='event', task__deadline__range=[start_date, end_date]) |
+                Q(task__type='ongoing', task__deadline__gte=start_date)
+            ).distinct()
+
+        project_type = query.get('project_type', None)
+        if project_type == 'volunteering':
+            qs = qs.annotate(Count('task')).filter(task__count__gt=0)
+        elif project_type == 'funding':
+            qs = qs.filter(amount_asked__gt=0)
+        elif project_type == 'voting':
+            qs = qs.filter(status__slug__in=['voting', 'voting-done'])
+
+        text = query.get('text', None)
+        if text:
+            qs = qs.filter(Q(title__icontains=text) |
+                           Q(location__name__icontains=text) |
+                           Q(pitch__icontains=text) |
+                           Q(description__icontains=text))
+
+        return self._ordering(query.get('ordering', None), qs, status)
+
+    def _ordering(self, ordering, queryset, status):
+        if ordering == 'deadline':
+            queryset = queryset.order_by('status', 'deadline', 'id')
+        elif ordering == 'amount_needed':
+            # Add the percentage that is still needed to the query and sort on that.
+            # This way we do not have to take currencies into account
+            queryset = queryset.annotate(percentage_needed=F('amount_needed') / (F('amount_asked') + 1))
+            queryset = queryset.order_by('status', 'percentage_needed', 'id')
+            queryset = queryset.filter(amount_needed__gt=0)
+        elif ordering == 'newest':
+            queryset = queryset.extra(
+                select={'has_campaign_started': 'campaign_started is null'})
+            queryset = queryset.order_by('status', 'has_campaign_started',
+                                         '-campaign_started', '-created', 'id')
+        elif ordering == 'popularity':
+            queryset = queryset.order_by('status', '-popularity', 'id')
+            if status == 5:
+                queryset = queryset.filter(amount_needed__gt=0)
+
+        elif ordering:
+            queryset = queryset.order_by('status', ordering)
+
+        return queryset
+
+
+class ProjectTinyPreviewList(ProjectListSearchMixin, OwnerListViewMixin, ListAPIView):
     queryset = Project.objects.all()
     pagination_class = TinyProjectPagination
     serializer_class = ProjectTinyPreviewSerializer
 
     owner_filter_field = 'owner'
 
+    permission_classes = (
+        OneOf(ResourcePermission, ResourceOwnerPermission),
+    )
+
     def get_queryset(self):
+        qs = super(ProjectTinyPreviewList, self).get_queryset()
         query = self.request.query_params
-        qs = Project.objects.search(query=query)
+        qs = self.search(qs, query=query)
         qs = qs.order_by('created')
         return qs.filter(status__viewable=True)
 
 
-class ProjectPreviewList(OwnerListViewMixin, ListAPIView):
+class ProjectPreviewList(ProjectListSearchMixin, OwnerListViewMixin, ListAPIView):
     queryset = Project.objects.all()
     pagination_class = ProjectPagination
     serializer_class = ProjectPreviewSerializer
 
     owner_filter_field = 'owner'
 
+    permission_classes = (
+        OneOf(ResourcePermission, ResourceOwnerPermission),
+    )
+
     def get_queryset(self):
+        qs = super(ProjectPreviewList, self).get_queryset()
         query = self.request.query_params
-        qs = Project.objects.search(query)
+        qs = self.search(qs, query)
         qs.select_related('task')
         return qs.filter(status__viewable=True)
 
@@ -60,6 +177,10 @@ class ProjectPreviewDetail(RetrieveAPIView):
     queryset = Project.objects.all()
     serializer_class = ProjectPreviewSerializer
     lookup_field = 'slug'
+
+    permission_classes = (
+        OneOf(ResourcePermission, ResourceOwnerPermission),
+    )
 
     def get_queryset(self):
         qs = super(ProjectPreviewDetail, self).get_queryset()
@@ -128,6 +249,10 @@ class ProjectDetail(RetrieveAPIView):
     queryset = Project.objects.all()
     serializer_class = ProjectSerializer
     lookup_field = 'slug'
+
+    permission_classes = (
+        OneOf(ResourcePermission, ResourceOwnerPermission),
+    )
 
 
 class ManageProjectPagination(BluebottlePagination):
