+++ bluebottle/members/models.py
@@ -108,5 +108,14 @@ class Member(BlueBottleBaseUser):
     def is_supporter(self):
         return self.amount_donated > 0
 
+    @property
+    def initials(self):
+        initials = ''
+        if self.first_name:
+            initials += self.first_name[0]
+        if self.last_name:
+            initials += self.last_name[0]
+
+        return self.first_name[0] + self.last_name[0]
 
 import signals # noqa
